WIP: skeleton of News-Events TabLayout
