+++ app/src/main/java/it/bz/noi/community/ui/today/NewsFragment.kt
@@ -1,0 +1,28 @@
+package it.bz.noi.community.ui.today
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import it.bz.noi.community.databinding.FragmentNewsBinding
+
+class NewsFragment  : Fragment() {
+
+	private var _binding: FragmentNewsBinding? = null
+	private val binding get() = _binding!!
+
+	override fun onCreateView(
+		inflater: LayoutInflater,
+		container: ViewGroup?,
+		savedInstanceState: Bundle?
+	): View {
+		_binding = FragmentNewsBinding.inflate(inflater, container, false)
+		return binding.root
+	}
+
+	override fun onDestroyView() {
+		super.onDestroyView()
+		_binding = null
+	}
+}
