+++ app/src/main/res/drawable/today_tabs_background_color_selector.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+	<item android:drawable="@color/primary_color" android:state_selected="true" />
+	<item android:drawable="@color/inactive2_color" />
+</selector>
