+++ app/src/main/res/values/themes.xml
@@ -52,4 +52,17 @@
 		<item name="cornerRadius">2dp</item>
 		<item name="backgroundTint">?attr/colorSecondary</item>
 	</style>
+
+	<style name="NOI.TabLayout" parent="Widget.MaterialComponents.TabLayout">
+		<item name="tabIndicatorHeight">0dp</item>
+		<item name="tabBackground">@drawable/today_tabs_background_color_selector</item>
+		<item name="tabTextAppearance">@style/NOI.TextAppearance.TabLayout.Button</item>
+	</style>
+
+	<style name="NOI.TextAppearance.TabLayout.Button" parent="TextAppearance.MaterialComponents.Button">
+		<item name="tabTextAppearance">@color/secondary_color</item>
+		<item name="tabSelectedTextColor">@color/secondary_color</item>
+		<item name="android:textAllCaps">false</item>
+		<item name="android:textSize">19sp</item>
+	</style>
 </resources>
